update: libdns 1.1.0
this commit finally implements the new libdns 1.x interface and types.

for this to work, new (public) methods ToLibDNS and ToAutoDNS are
introduced which wrap the logic of converting between AutoDNS types
and the individual resource record types of libdns.

with this commit, I am also adding an actual test suite which tries
to confirm the conversion methods work, and also adds a mock test to
confirm the request to the AutoDNS API endpoints works.

this does not (yet) test saving, but reading should be covered.

resolves: #9